Basics of scrolling properly implemented
The scrollbars now render as you would expect when you drag on the
canvas to move the display. We still need to implement mouse listeners
to allow you to grab the scrollbar widget and move the canvas. We still
need to tweak the rendering so we account for scrollbar heights and
widths to show the full canvas of alleles.