Start to make the rendering code multi-chromosome
This is the beginning of the process of making the rendering code multi-
chromosome aware. I've started with the map where a new class called
GenomeMap can be queried for a window of data in overall data
coordinates (i.e. within the total marker set). The return from this
method is an array of arrays, with each sub-array containing the marker
data for a given chromosome. I've also made the map rendering look more
like the desktop version of Flapjack by adding the rectangles which
represent a full chromosome.