Changes to load BrAPI v2 data
We've had to move away from BrAPI-js in the short term as the proposed
change to the pagination scheme would take too long to fully integrate
into the library at this point. This code will currently load Genotype
data from BrAPI a BrAPI variant set. It will also create a fake map for
the data as at this point the GOBii implementation doesn't have the
required markerpositions call.