Map rendering more closely matches desktop version
Some tweaks to the map rendering so that map rendering more closely
matches that of the desktop version. Once I have some time I want to
revisit this code as it needs to be simplified. The gaps between
chromosomes mean that too many markers are grabbed from the data model.