Line canvas bug fixes
The line names rendering code wasn't taking into account the possibility
of the scrollbar disappearing when the view is zoomed out. A similar fix
needed to be applied to the line name highlighting code to fix cases
where a line name would not render, but then be highlighted where the
scrollbar is theoretically being rendered over it.